Weird ClickOnce bugfix.
ClickOnce deployment was failing with "Deployment and application do not
have matching security zones.". Adding an UpdateUrl property seems to
fix this, despite doco suggesting that it's an unnecessary property when
the value is identical to InstallUrl. This also seems to have something
to do with the deploymentProvider element that gets generated (or not)
in the app manifest; seems we need it.